fix(plugins/filter/combine_lod): error on incomplete composite unique_key
A composite unique_key (list of keys) produced a tuple that is always
truthy, so an item missing one of the keys was silently grouped under a
(None, ...) key instead of raising. Validate each component explicitly.

Also adopt the standard Linuxfabrik file header, switch to f-strings,
fix the DOCUMENTATION so ansible-doc renders the filter again, and move
the embedded tests into tests/unit/plugins/filter/test_combine_lod.py
(plus new composite-key cases).
